samples/openai-gym/gym-cartpole-event-sample/requirements.txt
bonsai-ai>=2.0.10
bonsai-gym>=2.0.7
